requirements.txt
streamlit
PyPDF2
langchain
langchain_groq
chromadb
ollama
langchain_community
faiss-cpu
sentence-transformers
elevenlabs
groq
numpy
pygame
sounddevice
fastapi
multipart
wordninja
tavily-python